requirements.txt
absl-py==0.13.0
appnope==0.1.2
astunparse==1.6.3
attrs==21.2.0
autopep8==1.5.5
backcall==0.2.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
decorator==4.4.2
flake8==3.8.4
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h5py==2.10.0
idna==3.2
iniconfig==1.1.1
innvestigate==1.0.9
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
mccabe==0.6.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pluggy==0.13.1
prompt-toolkit==3.0.16
protobuf==3.17.3
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.8.0
pyparsing==2.4.7
pytest==6.2.4
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
tensorboard==2.2.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.57.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
